PORTAL-269 Portals should work without internet
Due to Google Visualization API, portals required internet to work

The loading of Visualization API has been moved close to the chart components so only them
will not be displayed in offline mode.
In scheduler, the charts in usage tab won't be displayed.
In RM, the monitoring tab will not be displayed as well as the charts on the bottom right

Change-Id: I21f0c04dc5f3995c4a51c8045b2853750b7cbfb6